# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KINECTRA API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /session/speech/synthesize:
    get:
      operationId: synthesizeSpeech
      tags: [session]
      summary: Synthesize text to speech
      parameters:
        - name: text
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Audio stream binary response
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
  /poses/search:
    post:
      operationId: searchPoses
      tags: [session]
      summary: Find matching professional pose
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PoseSearchInput"
      responses:
        "200":
          description: Matching professional pose details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoseMatch"
...
components:
  schemas:
...
    PoseMatch:
      type: object
      required:
        - matchName
        - similarity
        - role
        - description
        - idealVector
      properties:
        matchName:
          type: string
        similarity:
          type: number
        role:
          type: string
        description:
          type: string
        idealVector:
          type: array
          items:
            type: number
    PoseSearchInput:
      type: object
      required:
        - analysisType
        - poseVector
      properties:
        analysisType:
          type: string
          enum: [batting, bowling]
        poseVector:
          type: array
          items:
            type: number